# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  handlers.RecurringPaymentLink:
    properties:
      account_name:
        type: string
      amount:
        type: integer
      created_at:
        $ref: '#/definitions/models.Timestamp'
      description:
        type: string
      document_id:
        type: integer
      document_type:
        description: bill, invoice, payout or recurring_payment_occurrence
        type: string
      id:
        type: integer
      occurrence_due_date:
        type: string
      occurrence_status:
        type: string
      reference:
        type: string
      transaction_date:
        type: string
      transaction_id:
        type: integer
    type: object
  handlers.Response:
    properties:
      data: {}
      error:
        type: string
    type: object
...
  models.RecurringPayment:
    properties:
      account_id:
        type: integer
      account_name:
        description: Computed fields
        type: string
      amount:
        type: integer
      contact_id:
        type: integer
      contact_name:
        type: string
      created_at:
        $ref: '#/definitions/models.Timestamp'
      description:
        type: string
      end_date:
        type: string
      frequency:
        description: daily, weekly, monthly, quarterly, yearly
        type: string
      id:
        type: integer
      interval:
        description: every N frequencies (e.g. 2 = every 2 months)
        type: integer
      last_generated_date:
        type: string
      name:
        type: string
      next_due_date:
        type: string
      reference:
        type: string
      start_date:
        type: string
      status:
        description: active, paused, cancelled, completed
        type: string
      type:
        description: income, expense
        type: string
      updated_at:
        $ref: '#/definitions/models.Timestamp'
    type: object
  models.RecurringPaymentInput:
    properties:
      account_id:
        type: integer
      amount:
        type: integer
      contact_id:
        type: integer
      description:
        type: string
      end_date:
        type: string
      frequency:
        type: string
      interval:
        type: integer
      last_generated_date:
        type: string
      name:
        type: string
      next_due_date:
        type: string
      reference:
        type: string
      start_date:
        type: string
      status:
        type: string
      type:
        type: string
    type: object
  models.RecurringPaymentOccurrence:
    properties:
      allocated:
        description: Computed fields
        type: integer
      amount:
        type: integer
      created_at:
        $ref: '#/definitions/models.Timestamp'
      due_date:
        type: string
      id:
        type: integer
      payment_name:
        type: string
      recurring_payment_id:
        type: integer
      status:
        description: pending, paid, skipped
        type: string
      unallocated:
        type: integer
      updated_at:
        $ref: '#/definitions/models.Timestamp'
    type: object
  models.Timestamp:
    properties:
      time.Time:
        type: string
    type: object
...
host: localhost:8090
info:
  contact: {}
  description: API for managing accounts, contacts, bills, invoices, and transactions.
  title: Portal API
  version: 1.0.0
paths:
...
  /recurring-payments:
    get:
      description: Get a list of all scheduled recurring payments (income or expense).
      parameters:
      - description: Filter by status (active, paused, cancelled, completed)
        in: query
        name: status
        type: string
      - description: Filter by account
        in: query
        name: account_id
        type: integer
      - description: Filter by type (income, expense)
        in: query
        name: type
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/handlers.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.RecurringPayment'
                  type: array
              type: object
      security:
      - BearerAuth: []
      summary: List recurring payments
      tags:
      - recurring_payments
    post:
      consumes:
      - application/json
      description: Create a new scheduled recurring payment.
      parameters:
      - description: Recurring payment details
        in: body
        name: recurring_payment
        required: true
        schema:
          $ref: '#/definitions/models.RecurringPaymentInput'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/handlers.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.RecurringPayment'
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/handlers.Response'
            - properties:
                error:
                  type: string
              type: object
      security:
      - BearerAuth: []
      summary: Create recurring payment
      tags:
      - recurring_payments
  /recurring-payments/{id}:
    delete:
      description: Remove a recurring payment.
      parameters:
      - description: Recurring Payment ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/handlers.Response'
            - properties:
                data:
                  additionalProperties:
                    type: string
                  type: object
              type: object
        "404":
          description: Not Found
          schema:
            allOf:
            - $ref: '#/definitions/handlers.Response'
            - properties:
                error:
                  type: string
              type: object
      security:
      - BearerAuth: []
      summary: Delete recurring payment
      tags:
      - recurring_payments
    get:
      description: Get details of a specific recurring payment.
      parameters:
      - description: Recurring Payment ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/handlers.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.RecurringPayment'
              type: object
        "404":
          description: Not Found
          schema:
            allOf:
            - $ref: '#/definitions/handlers.Response'
            - properties:
                error:
                  type: string
              type: object
      security:
      - BearerAuth: []
      summary: Get recurring payment
      tags:
      - recurring_payments
    put:
      consumes:
      - application/json
      description: Update details of an existing recurring payment.
      parameters:
      - description: Recurring Payment ID
        in: path
        name: id
        required: true
        type: integer
      - description: Updated recurring payment details
        in: body
        name: recurring_payment
        required: true
        schema:
          $ref: '#/definitions/models.RecurringPaymentInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/handlers.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.RecurringPayment'
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/handlers.Response'
            - properties:
                error:
                  type: string
              type: object
        "404":
          description: Not Found
          schema:
            allOf:
            - $ref: '#/definitions/handlers.Response'
            - properties:
                error:
                  type: string
              type: object
      security:
      - BearerAuth: []
      summary: Update recurring payment
      tags:
      - recurring_payments
  /recurring-payments/{id}/links:
    get:
      description: Get all payment transactions linked to a specific recurring payment,
        including the due date and status of the occurrence each transaction covers.
      parameters:
      - description: Recurring Payment ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/handlers.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/handlers.RecurringPaymentLink'
                  type: array
              type: object
      security:
      - BearerAuth: []
      summary: Get recurring payment links
      tags:
      - recurring_payments
...
  /recurring-payments/{id}/occurrences:
    get:
      description: Returns all scheduled occurrences (instances) for a recurring payment.
        Occurrences are auto-generated by the server for each due date up to today.
        Each occurrence tracks whether it has been paid via a linked bank transaction.
      parameters:
      - description: Recurring Payment ID
        in: path
        name: id
        required: true
        type: integer
      - description: Filter by status (pending, paid, skipped)
        in: query
        name: status
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/handlers.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.RecurringPaymentOccurrence'
                  type: array
              type: object
        "404":
          description: Not Found
          schema:
            allOf:
            - $ref: '#/definitions/handlers.Response'
            - properties:
                error:
                  type: string
              type: object
      security:
      - BearerAuth: []
      summary: List recurring payment occurrences
      tags:
      - recurring_payments
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"